Revert "refactor+test: decouple bi-diagram getFlowModel for headless testing"
This reverts commit 6a7526163733496c4621c3487dd7ea543cf01325.
